fix (?): Remove 'aria-roledescription' from map
Lighthouse seems to have a bug that prevents it from recognizing
'aria-roledescription' as a valid WAI-ARIA attribute.  The
Lighthouse development team is aware of the issue but to date they
haven't fixed it.  So, for the purpose of meeting my Lighthouse
accessibility percentage for this assignment, I am removing the
attribute.