chore: Rewrite gulpfile.js so that I can understand it
I made the mistake of using Yeoman to generate my gulpfile without
really understanding how Yeoman or Gulp work, and this prevented
from from understanding why certain errors were happening with my
build and how I could address them.

Re-writing the gulpfile helped with this as I now now have a better
understanding of gulp and babel and their various ecosystems and
the build script is cleaner as well.

I've commented out the linting and uglifying steps because they're
not compatible in their current configuration with the Javascript
I'm writing.  I hope to rectify that as time goes on.